Minje Week 24

+++ MinJe/Week 23/6198_옥상정원꾸미기.cpp
@@ -10,7 +10,7 @@ int main(){
     cin >> n;
 
     stack<int> buildings;
-    long long answer = 0;
+    unsigned int answer = 0;
 
     for (int i = 0; i < n; i++){
         int height;
